translating literal strings shouldn't be picked up as strings
for whatever reason, Django's TranslateNode converts string literals
into a variable, both in 1.8 and 1.9

this probably should be reported upstream, but for now, we work around
it. Had to refactor _get_vars(filter_expression) so that results is
always initialized (preferred this to adding an extra else block just to
initialize it empty.

used explicitly named variables becuase it helped me in debugging, so
didn't inline them back

Now _get_vars can be forced not to include the basic variables, only the filter ones